feat(devex): Introduce new `prettier-plugin-tailwindcss` plugin
Tailwind classes dont *need* to be sorted, but it's a great idea to follow some kind of standard. Let's add this plugin to keep them sorted, we might develop some muscle memory and will start writing in the "right" way as well. One of the good things about this is that it sorts them in the order they're applied in the browser, which means it's easy to reason about how classes override one another.