feat(revenue-analytics): Add new/churned customer metrics to query
This was supposed to be here since the first day, but adding these was slightly trickier than expected. We now have "New Customer Count" and "Churned Customer Count" on our dashboard. I tried several different approaches for this, involving WINDOW queries but in the end I settled on a simple approach where I'm just querying the current AND previous at the same time in Clickhouse, this is fast enough for now.