feat(revenue-analytics): Add Revenue Analytics ARPU query and associated components
Implemented `RevenueAnalyticsArpuQueryRunner` for executing the ARPU query. This simply involves calculating *all* revenue stemming from the `InvoiceItem` view and then dividing by the amount of active users in that period. It might make sense to change this for MRR eventually.

One point for improvement here is that if we could simply reuse the queries from `MRR` and `customer_count` then this would run almost instantaneously! We might wanna explore creating a tree of queries here, and feed them to the UI in a simple way. Pre-aggregation might work too.

Closes #34005