devex: Add prettier plugin to sort import
We've recently moved from eslint to oxlint which caused us to lose our lints that would sort our imports. Let's introduce a new prettier plugin (prettier-plugin-sort-imports) to compensate for that. This plugin is highly customizable, and I tried to make sure the rules it's following align closely to what we had before. There's still a LOT of changes to it because they add line breaks where we didn't add before, hence the 2.4k files changed.