feat(revenue-analytics): implement cache target age logic for revenue data syncing
We're hitting a major problem with Revenue Analytics that's hurting our onboarding process. Users add their first revenue source to PostHog and then open the Revenue Analytics dashboard. We'll then query their partial data and display the results, but we also display a big banner saying that we're still syncing their data so it's still in progress. This is all fine.

What happens, however, is that people come back to the UI after 30 minutes (enough time to sync medium-sized accounts) and the banner is gone, but the charts are still displaying stale data because we've cached all those insights. They can just click the `refresh` button, but that's not obvious.

To reduce that problem, let's be much smarter with our caching mechanism:
- when we detect we're syncing data from a revenue source, let's implement a minimal cache (just 1min to avoid flooding)
- when not syncing, rather than using the default caching period, let's look at the sync frequency for the sources and use half the shortest sync period as our cache

This will guarantee that customers should see non-stale data very frequently. We're, of course, sacrificing speed/CPU for this, but I believe it's acceptable given the increase in user satisfaction.

If we ever think this is hurting our CPU, we can find smarter ways to bust the query - for example, clear the cache when DWH syncing finishes via some kind of hook (that doesnt exist yet).

Closes #34650