feat: Add product/coupon properties to RA event sources
Revenue Analytics simply ignored product/coupons for event sources - let's expand our configuration to allow for that.

We're adding two new properties to our configuration table, letting customers customize how they can identify a product and their cupon. This is fully supported on filters too.

The only cons is that we can't properly handle `initial_coupon` because we aren't setting that property on persons. We can handle this in the future if needed, but queries will likely be slower - options to implement this are: join from persons to events to figure out earliest presence of `coupon` (slow), or instruct people to `$set_once`, or maybe use materialized virtual properties.

See https://posthog.com/questions/product-breakdow