feat(devex): Bump prettier and kea-typegen
We wanna add a new prettier plugin to properly sort imports (https://github.com/trivago/prettier-plugin-sort-imports). This will require kea-typegen v3.5+ and Prettier v3+. Let's prepare the ground by upgrading prettier/kea first and running it against the whole codebase.

There's a small problem, though: oxlint had to run in the whole codebase because of all these changes which caused it to change a lot of our `useEffect`s to add missing dependencies. This is very dangerous, so I've:
- created a new `useEffectOnMount` helper that hardcodes an empty array
- used the helper where we had an empty list of dependencies
- ignored the changes wherever oxlint WOULD have changed

This guarantees the behavior is the same as the one we have now. There are several places where we SHOULD add the dependencies to the list, but this would be too hard to test, so let's keep this as is for now 