feat(revenue): make revenue analytics work for poe
This is not 100% what I want because the `revenue_last_30_days` takes the query time into consideration rather than the person at the time of the event, but that's fairly hard to do in Clickhouse.

In PostgreSQL we'd do something akin to (pseudosql)

```sql
SELECT
  id,
  (SELECT sum(revenue) FROM invoices WHERE person_id = events.person_id AND timestamp IN (events.timestamp - 30 days, events.timestamp))
FROM events
```

Postgres will be smart enough to only join this once and run the filters on each subquery in a smart way. Clickhouse doesnt work that way. I cant find an easy way to do this after working on this stuff for well over 30 hours now, so... will only come back once someone asks me for it as a feature.

Closes #35973