Revert changes to `remote_config` endpoint
I made some changes to the `remote_config` endpoint to try and make project selection deterministic. It turns out there's already a supported and simpler way to do it: pass the project api token in the `token` query string parameter.

This is what the `local_evaluation` endpoint does, so we should follow that example.

I am not reverting the changes to `ProjectSecretAPIKeyAuthentication` because those are actually needed.