Fix `send_cohort` behavior when `use_cache` is set
The problem is `send_cohorts` is a little confusing. We always send cohorts as part of the response, but if `send_cohorts=true` we send cohort definitions in a top-level cohorts field.

If `send_cohorts=false`, we transform the filters for the flags that have cohort filters to include the cohort conditions. But this only supports simple person property filters.