refactor: Extract `ExternalDataSourceType` class
We've had a lot of problems since we started using `generated_configs.py` because it's very prone to generating import loops since it depended on `posthog.warehouse.models.ExternalDataSource`. Moving `ExternalDataSource.Type` to `posthog.warehouse.types.ExternalDataSourceType` makes that problem much less apparent, so here it is! This is also a better practice as well, less nesting makes importing in Python much easier overall.

I've also gone ahead and updated the `schema` file to fix that weird `Type4` we had in the generated config, so consider that a bonus :)