feat(revenue-analytics): add team_id to `persons_revenue_analytics` table
Introduced `team_id` alias in the `select_from_persons_revenue_analytics_table` function to ensure compatibility with HogQL (always adds `team_id` clause to queries).

This is one of the only "tables" (it's lazy) in PostHog that don't include the `team_id` so it could be considered a security risk. It is NOT a security risk in practice because it is created from Revenue Analytics views, which are themselves created from DWH views and the events table, which are safe.

I've also renamed it to `persons_revenue_analytics` to better represent what it is. I'd like to eventually call it just `revenue_analytics` but that's not possible right now because a `TableGroup` cannot contain a leaf table in it, therefore it means we cannot have both `revenue_analytics` AND `revenue_analytics.events.purchase` at the same time - this should be fixed eventually.

Closes #36687