feat: Return new/expansion/contraction/churn revenue to the UI
The query runner now knows how to compute all of this. We're reusing `RevenueQueryRunner` because it already handled computing most of it and updating it was fairly simple - it's also faster than recomputing this data all the time.

The basic idea here is that we're grouping the revenue per subscription and then computing all of these in respect to a single subscription. For subscriptionless events we're simply assuming they're each their individual subscription, adding up to both `new` and `churn`ed revenue at the same time.

We might need to change our implementation if we ever decide to show "who churned", "who grew", in a global table. It will be much easier to achieve once we have a pipeline rather than simply computing these during query time.

We do NOT have UI for this yet, this is coming in a future PR.

See #34007