refactor: Remove inheritance from structlog.WriteLogger
Previous implementation was breaking Liskov substitution principle as
NamedLogger subclasses could not be used in place of
strucltog.WriteLogger.

So, this commit prefers composition over inheritance by holding a
WriteLogger as an attribute of Logger. Write calls are then dispached
to the underlying logger.

This removes the need for a NamedLogger class entirely, and removes
some mypy comments.