feat: add goal lines support to trends funnel queries
I'm working with a friend today and he's a PostHog user. He said he'd love to have goal lines on trend funnels (for example, he wants conversion to hit 2%) so I've implemented it in 20min!

Very simple implementation, just copying over a lot of the implementation we already had for Trends