feat: Implement improved MRR query
We're now calculating the MRR query entirely on Clickhouse. Previously we'd try and attempt to compute what "MRR" meant by doing that at the Python level, but that made it very tricky to calculate the "initial" amount and what the MRR for a given "month" was. We're now solving that by creating a virtual view that includes the MRR for a given customer on each day, and then assume the last day of the month is what counts as the "month" MRR when we're filtering by month

This is slightly heavier on Clickhouse, but it's still pretty quick!

This implies the existing "MRR Growth" chart is doing the wrong calculation, I'm gonna remove it soon and improve on this one instead