fix(flags): Use token-based key for local evaluation hypercache
The `local_evaluation` endpoint is called by the SDKs which use the project api key to identify the project.

Since the SDKs don't know the team id, we need to support cache keys using the project api token, otherwise we'll need a database lookup defeating the purpose of this caching work.