refactor: Move tag label over the ellipsis dropdown selector
Revenue Analytics has a very long name and its `beta` label is hidden behind the `truncate` we have on the sidebar. To circumvent that, let's inside render labels over the `...` ellipsis we use as dropdown selector.

Of course, we display the ellipsis on hover with some CSS `group` magic. I'm still in doubt on whether we should center or right-align the ellipsis with the badge, so I'll leave that to the Design gods.

https://www.loom.com/share/5497d7a2048f4ac3a7f87ade2b6bbc4c