refactor: Big refactor on how we compute breakdown
We're now properly using taxonomy to compute breakdown rather than just creating a custom thing. This is much nicer! We had to rework a lot of stuff, but the code is easier to grasp now. We still need to update the UI.

We've also removed the "growth" component in this PR since we can introduce it back using frontend-only stuff (trends introduced by Prod Analytics team)