fix: Use proper field for data warehouse schema default incremental field
We use this field in the UI to match against the available "incremental fields" and save the table. We match against the `value` field though, which meant the UI was displaying the right value (right label) but it required the user to click on the select and reselect the field to get the actual value. This wasn't a problem for most integrations since the field/label pair is usually the same, but it affected Stripe more heavily since the field/label are `created`/`created_at`.