fix(flags): Prevent pending conditions from moving
This is a naive fix to address an edge case where a user could modify
conditions, causing the affected user count to be recalculated, and then
quickly move the condition up or down. When the API response finally
returns, it would assign the affected user count to the incorrect index
(as it's now been moved).

This change prevents any conditions pending affected user calculations
from being moved until they resolve.