feat: Allow `SelectViewType` types field resolution
Right now we cannot resolve the proper type of view fields since we're only resolving it for `BaseTableType`. This is very restrictive and not very useful since we store type information for all saved views - and Revenue Analytics managed views - in the database and, therefore, have easy and accurate type information.

This is easy to see in production with the Revenue Analytics views where we have some JSON fields - such as `metadata`. For JSON fields in a real "table" like the event `properties` we can easily access it as `properties.key` and both the SQL editor and the actual query properly resolve that. If we do the same for `metadata.field` then it fails because it does not know that `metadata` is a JSON field and doesn't apply the proper `extractRawJSON*` queries.