feat: Add reverse lookup for Cohorts
This change introduces a new table, `posthog_cohortdependency`, which
maps and indexes dependencies between cohorts for fast lookups. This
will allow us to quickly identify which cohorts must be recalculated
should a cohort be updated. Previously, this was only feasible when
traversing up the dependency chain. E.g., Cohort A depends on Cohort B,
Cohort A updates, so should Cohort B. This is possible because we can
follow the dependency chain when we update Cohort A.

However, what if the opposite is true, and Cohort B depends on Cohort A?
When Cohort A is recalculated, Cohort B must also recalculate, but we'd
need to iterate every cohort within the project and evaluate its
properties to identify the dependency chain.

With this change, it's a simple indexed lookup into PostgreSQL. As
records within `posthog_cohort` are created/updated/deleted, triggers
will maintain the persistance of the index in
`posthog_cohortdependency`.