chore(flags): Remove `insert_in_clickhouse` argument to `insert_users_list_by_uuid`
In 99.9% of cases, when we call `insert_users_list_by_uuid` we want to also insert them into Clickhouse (aka `insert_in_clickhouse=True`). The only time we don't want to do that is when calling `insert_cohort_from_insight_filter` or `insert_cohort_from_query`. And the reason for that is those methods create the cohort in Clickhouse first, and then populate Postgres with the IDs from Clickhouse.

Unfortunately, `insert_in_clickhouse` is defaulted to `False`. This means it's very easy to introduce a bug if not careful.

To avoid that, this PR removes the `insert_in_clickhouse` argument from `insert_users_list_by_uuid` and changes the method to always insert into Clickhouse.

It also adds a new `insert_users_list_by_uuid_into_pg_only` method that those other methods can leverage.