fix: Keep Rev Analytics onboarding state active while not finished
We were previously automatically leaving the Revenue Analytics onboarding as soon as we connected a source/event but that's not ideal since the user might not have finished configuring Revenue Analytics yet.

Instead, let's now "block" the user inside the onboarding until we actually finish onboarding and the user clicks "view dashboard".