fix: Support joins with tables that include `timestamp`
By mistake we were not supporting filters on tables that also include a `timestamp` field - because we were not specifying what `timestamp` field we were filtering by. This didn't show up locally because our queries got cached... I wanna solve that so bad, it's kinda wild that this is a thing.

I've now added a test to prevent such a regression from happening.