feat: Expose Revenue Analytics insights/queries to Max
That's a lot of work, but Max has access to the Revenue analytics queries in its context now!

Followed the README.md file and updated it slightly because there was a tiny thing missing.

Also, split the `format.py` and `test_format.py` file into several files, should be much more organized now