feat: Add Revenue analytics filters to Max AI
We can now customize dates, filters and breakdowns in Revenue Analytics. This was implemented by copying what we already had for Session Replay and expanding the taxonomy agent slightly so that it can learn how to find the relevant revenue analytics properties.

Local tests seem to show this works well, but I'd appreciate some help figuring out whether I should write tests for this or not