feat(revenue-analytics): alter DataWarehouseSavedQuery model to include query_type and external_data_source fields
We wanna be able to define a query from the backend and not have it stored in the DB. To make that the simplest possible let's hook on the `query` method and return our stored query when needed. `get_columns` was also changed because we don't need to go all the way to the DB to figure out types since we have that defined in code already.

Serializer is returning that field to the UI, just not using it yet.