feat: Support revenue analytics insights in the UI
This is not the prettiest code, but because revenue analytics insights are not really insights we need to find a way to access them via the `revenueAnalyticsLogic` rather than through the `insightsLogic` we already have.

This does the trick in the simplest way possible while still maintaining the impression to the backend that this is an insight.