feat(flags): Notify when a flag uses unsupported regex
Flags which may be resolved remotely via the flags service cannot use
look-ahead, look-behind or back-references. The `regex` crate does not
support these.

The number of unsupported regex features is small enough that
client-side validation is preferred. Otherwise, we'd likely need to
persist a value on the flag indicating whether or not the flags service
can parse the regex for user feedback.