refactor: Return single `BuiltQuery` rather than Iterable
This makes our code much simpler. Previously we'd return an iterable of `BuiltQuer`ies if we were using events - one for each of our events - but for DWH integrations we'd always return only one item in the list.

There's no reason for that to work that way. We can bring them both to always returning a single query by hoisting the "one view per event" generation to the `orchestrator.py` in the same way we have it for DWH sources.

This both simplifies the view generation but also our tests, win-win.