feat: Introduce concept of managed view
We wanna move our revenue analytics queries to saved views because we want them to be materialized. To achieve that, we're introducing the concept of managed views!

These have a specific type and they're backed by a `SavedView` to handle materialization for us. For now, we're not returning these saved views to the UI or anything, this PR/commit is simply meant to identify whether the data modeling makes sense or it should work differently.

In the future we might have non-revenue analytics-related managed views - the architecture easily supports that.

We are also not handling enabling/disabling revenue analytics sources - again, this will come in a follow-up PR, the current code is only meant as means to review the architecture. This might get broken up in many smaller if needed.