fix(flags): Maintain blast radius associations
When moving a condition set as a blast radius is calculating, the
affected user count will no longer be assigned to the old index. This
would cause numbers for a given condition set to be incorrect.

This change decouples the affected user counts from the index, instead
opting to use the stable sort_key.