refactor: Fix churn calculation for revenue analytics
Previously we'd count the churn on the month the customer's subscription ended. This caused a small problem because we'd reflect new subscriptions/customers immediately on the same month/day it started, but we'd only decrease it from the active subscription/customer count the next day (because the subscription was active on that day!)

This meant that when looking at a chart of active/new/churned subscriptions that's how it'd look like:

```
Active:  10  11  11  11  10  10
New:      0   1   0   0   0   0
Churned:  0   0   0   1   0   0
```

You can see how someone churns on the 4th day, but it's only decreased from active subscriptions on the next day. I believe the chart should look like as follow, which prompted the change here

```
Active:  10  11  11  11  10  10
New:      0   1   0   0   0   0
Churned:  0   0   0   0   1   0
```

See the last row, it's now only churned on the 5th day, following the decrease on the first row on the 5th day.

This could also be solved in a different way, by decreasing the number of active subscriptions on the 4th day. This could easily be achieved if Clickhouse included a `toEndOfPeriod` function but alas it doesn't.

I'm checking with @minekansu whether this change makes any sense or not