feat: Add `sortableSemver` HogQL function
This function properly sorts Semver items in a reasonable order rather than simple alphabetical order, i.e. it properly ranks 1.10.0 above both 1.1.0 and 1.2.0 (and 1.9.213323232 even).

This is used in the SDK Doctor only for now, but we could expand to use it everywhere.

Also moved some functions around to a better place