refactor: Standardize HogQL functions
Most of our HogQL functions are following `camelCase` with the exception of a handful of them, so let's standardize them all! This is a breaking change but I don't know how to check whether this is gonna break something yet, might need some help with this.

Also, I've removed the `hogql_` prefix from the channel type ones since we don't have that prefix anywhere else