feat(managed-viewsets): add managed viewsets functionality with UI components and API integration
Introduced new ManagedViewset in the backend to group/manage several views together. Right now this will be used by revenue analytics to create actual saved queries rather than simply having virtual ones.

We are already creating the views for those who have the FF enabled, but are not using them anywhere - they're filtered out both in the UI and the `create_hogql_database` helper

One of our concerns was: how do we resync them when the underlying view structure changes and that's been solved too! The function that creates/updates these views overrides the query/columns when syncing them.

We've added it to the UI too to make it easier for me to test the changes/resync it in prod, and there's even a new onboarding flow for revenue analytics that depends on that