tests: Remove incorrect test
The assumptions behind this test are wrong. We'll never have two saved views without any paths that will have their paths created after the fact. A view is always created with a path attached to it.

It's possible we update a view and fall into this trap, but this is tested below.

Follow-up to ad3924cd2aecd4ee6dfc4a16dcad7667e95ccd90