feat(flags): add 5-second default for reader statement_timeout settings
Set default statement_timeout to 5000ms (5 seconds) for the reader pools. Set it to 1000ms (10 seconds) for the writer.
This helps prevent runaway queries by default.

Also, when the DB is under heavy load, it allows long running queries to be canceled to return the connection to run other queries.