feat: Use managed viewset views for revenue analytics
This adds all the code necessary to display the managed viewset views in the UI and actually use them versus using the ones we were generating on the fly. It'll properly choose to use the materialized views if they exist which is very cool!

All hidden behind a FF, very easy to rollback