Ensure missing `project_id` doesn't break deserialization
Noticed a lot of `failed to parse data to team for token phc_...` errors in our logs. This gets treated as a cache miss so the `/flags` service ends up querying the database and writing to the cache.

This adds a `#[serde(default)]` to the `project_id` field in the `Team` struct so it defaults to `0`. `Team::from_redis` ensures `project_id` is set to `team.id` if the value is `0`.