Implement a generic `ReadThroughCache`
We have a lot of places where we attempt to read a Redis cache entry and if the entry doesn't exist, we fetch the data from the database and then write the entry to a writer `RedisClient`.

The new `ReadThroughCache` class encapsulates and improves on that pattern.

For example, our existing code:

1. Treats `TimeoutError` and other errors as cache misses, causing us to attempt to write to the cache when it might be under load.
2. Doesn't evict existing entries when the DB fetch produces no results.
3. Doesn't distinguish between corrupt data and missing data.
4. Doesn't cache negative entries. So if a project api key doesn't exist, requests with it always hit the db.

The `ReadThroughCache` (combined with `NegativeCache`) provides the structure to fix all these issues.