feat: Support session replay classes on toolbar PII tool
Let's support more stuff here not just the basic `ph-no-capture`. It's slightly harder to support RegExp/Javascript functions since it'd require a MutationObserver which might be a bit painful to set up (and this is just a hackhathon project).

In those cases, we simply display a warning telling customers we don't support that in the tool.