fix: show warning when event was buffering (#41423)
when putting a view recording button in the page we pass a status from the event that we're linking from

if the status is buffering we show the button as enabled

this makes sense since we don't know there will be no recording

but

it may be that session is buffering waiting for a trigger that it never sees, and so you click through and get not found

confusion!
despair!
bad UX!

the button already has the concept of a warning

so, let's add a warning that the recording was buffering and _might_ be watchable

![2025-11-13 11.02.50.gif](https://app.graphite.com/user-attachments/assets/e34d28cd-c028-43e6-a750-decc407bfec7.gif)