refactor: Standardize materialization even further
We'd created this `enable_materialization` method to centralize our logic but we didn't use it for endpoints, so let's use that now! I've also changed it to `schedule_materialization` since this is a more accurate name (you still need to configure materialization settings before calling it).

There's also a fix in `create_leaf_paths_from_query` to guarantee we follow the same logic we had in `recreate_model_paths` which was only called when updating a saved query (which was called when turning materialization on and, therefore, a step that was needed and we missed on the previous attempt to standardize this). This meant that revenue analytics and endpoints materialized queries would fail in case they were simple constant queries (i.e. no `FROM` field)