feat: Generate `UserProductList` when we detect product intent
One of the cases where we wanna automatically generate a new `UserProductList` is when we
detect product intent - which teams already trigger from the UI.

This just hooks onto `ProductIntent.register` and creates the item in the sidebar. The UI
is also automatically refreshing to show it on the sidebar - properly respecting user preferences, ofc.