fix(temporal-worker): Graceful shutdown of activity executor
We are not gracefully shutting down the
concurrent.futures.ThreadPoolExecutor used in the Temporal worker.
Maybe Temporal is doing this for us, but it's not clear.

We should manage it's lifecycle by gracefully shutting down using a
context manager.

This passes a concurrent.futures.ThreadPoolExecutor to the Worker.
This executor is managed with a context manager, so upon exiting the
asyncio.Runner we will shut the executor down next.

Log lines added on every shutdown step.