feat: Paginate Github SDK fetching
We are now properly paginating across all the github pages to get all the available releases.
This means we'll not get only the last 4 or 5 `posthog-js` SDKs we were getting - we release too fast.

We've also moved all of the SDKs to use the `Releases` method. This required backfilling releases and updating
SDKs jobs to guarantee they always have a release. This has been done already.