chore: Implement a update_billing_as_actor method in billing_manager.py and use it for update_
When a user leaves/enters an organization it attempts to update the users of the organization on the
billing side. Until now it used the user itself, however that user may not have the proper permissions
to call /api/billing. I've implemented a update_billing_as_actor method that authorizes using a different
user. Kind of like sudo.